Refactorizar rutas de imágenes en estilos CSS y componentes
