+++ public/locales/en/translation.json
@@ -44,7 +44,7 @@
   "emailNotValid": "The email is not valid",
   "emailPlaceholder": "Email",
   "emailSupport": "support@gopass.com",
-  "enterCode": "Enter the code",
+  "enterCode": "You will shortly receive an SMS with a verification code.",
   "enterEmail": "Enter your email",
   "entries": "Entries",
   "entryDetails": "Entry Details",
@@ -57,6 +57,7 @@
   "forgotPassword": "Forgot your password?",
   "fullName": "Full Name",
   "gopassComimssion": "GOPASS Commission (8%):",
+  "here": "Here",
   "help": "Help & Support",
   "help&SupportGP": "Help & Support - GoPass",
   "howDoesTicket...": "How does ticket verification work on GoPass?",
@@ -156,6 +157,7 @@
   "verifyStep1": "Log in to your GoPass account.",
   "verifyStep2": "Enter the alphanumeric code found below the QR code.",
   "verifyStep3": "Wait for the system to verify the authenticity of your ticket and you're done.",
+  "veriyProfile": "Complete the profile data, and you will be able to verify your phone number to buy tickets.",
   "viewMore": "View more",
   "waitForTheSystem...": "Wait for the system to complete the verification (this may take up to 24 hours).",
   "welcomeText": "Welcome to GoPass, your reliable solution for ticket verification.",
