+++ public/locales/es/translation.json
@@ -44,7 +44,7 @@
   "emailNotValid": "El correo electrónico no es válido",
   "emailPlaceholder": "Correo Electrónico",
   "emailSupport": "soporte@gopass.com",
-  "enterCode": "Introduce el código",
+  "enterCode": "Recibirás un sms con un código de verificación.",
   "enterEmail": "Introduce tu correo electrónico",
   "entries": "Entradas",
   "entryDetails": "Detalles de la entrada",
@@ -64,6 +64,7 @@
   "howToBuy": "¿Cómo comprar una entrada en GoPass?",
   "howToSell...": "¿Cómo revender una entrada a través de GoPass?",
   "howToVerify...": "¿Cómo verificar una entrada en GoPass?",
+  "here": "Aquí",
   "image": "Imagen",
   "IUnderstandThatIAm...": "Entiendo que estoy adquiriendo una entrada a través de un tercero en reventa y que la plataforma garantiza la autenticidad de la entrada en caso de que sea validada correctamente.",
   "latestTickets": "Últimas entradas",
@@ -156,6 +157,7 @@
   "verifyStep1": "Inicia sesión en tu cuenta de GoPass.",
   "verifyStep2": "Introduce el código alfanumérico que se encuentra debajo del código QR.",
   "verifyStep3": "Espera a que el sistema verifique la autenticidad de tu entrada y listo.",
+  "veriyProfile": "Completa los datos del perfil, y podrás verificar tu número de teléfono para comprar entradas.",
   "viewMore": "Ver más",
   "waitForTheSystem...": "Espera a que el sistema complete la verificación (esto puede tardar hasta 24 horas).",
   "welcomeText": "Bienvenido a GoPass, tu solución confiable para la verificación de entradas.",
