+++ src/app/pages/MyTicketsPage.tsx
@@ -28,7 +28,7 @@ export const MyTickets: React.FC = () => {
         <section className="flex w-full flex-col items-center">
           {tickets.length === 0 ? (
             <div className="absolute flex h-full w-full items-center justify-center">
-              <p className="text-xl font-semibold text-gray-500">No hay entradas disponibles</p>
+              <p className="text-xl font-semibold text-gray-500">{t("noEntriesAvailable")}</p>
             </div>
           ) : (
             <div className="mt-20 w-[90%]">
