+++ src/app/pages/PhoneCodePage.tsx
@@ -34,7 +34,7 @@ export default function ForgotPassword() {
       <img className="absolute -z-50 h-screen w-full" src="/photo-1674307419403-9f4991ce4ce1.avif" />
 
       <div className="flex h-screen flex-col items-center justify-center">
-        <p className="mb-6 text-3xl font-black text-customWhite">Introduce el código recibido.</p>
+        <p className="mb-6 text-3xl font-black text-customWhite">{t("enterCode")}.</p>
 
         <form onSubmit={handlePhoneCode} className="flex w-[90%] flex-col gap-5 rounded-2xl bg-[#e0e0e0e2] p-4">
           <InputField
