+++ src/hooks/usePhoneCode.ts
@@ -1,5 +1,6 @@
 import { SystemError } from "com/errors"
 import { useNavigate } from "react-router-dom"
+import Swal from "sweetalert2"
 
 import { phoneCodeFetch } from "../service/phoneCode"
 
@@ -8,7 +9,16 @@ export const usePhoneCode = () => {
   const phoneCodeCorrect = async (vonageCode: number): Promise<void> => {
     try {
       await phoneCodeFetch(vonageCode)
-      navigate("/login")
+      Swal.fire({
+        title: "Éxito",
+        text: "Número de teléfono verificado",
+        icon: "success",
+        confirmButtonText: "Aceptar",
+      }).then((result) => {
+        if (result.isConfirmed) {
+          navigate("/login")
+        }
+      })
     } catch (error: any) {
       if (error instanceof SystemError) {
         alert(error.message)
