+++ src/privateRoutes/PrivateRoutes.tsx
@@ -1,17 +1,17 @@
 import React from "react"
 import { Route, Routes } from "react-router-dom"
 
-import { AboutUs } from "../app/pages/AboutUs"
-import AllTickets from "../app/pages/AllTickets"
+import { AboutUs } from "../app/pages/AboutUsPage"
+import AllTickets from "../app/pages/AllTicketsPage"
 import ComprarEntrada from "../app/pages/BuyEntryPage"
 import Contact from "../app/pages/ContactPage"
-import { HelpSupport } from "../app/pages/HelpSupport"
+import { HelpSupport } from "../app/pages/HelpSupportPage"
 import LandingView from "../app/pages/LandingViewPage"
-import { MyTickets } from "../app/pages/MyTickets"
-import { MyTicketsSold } from "../app/pages/MyTicketsSold"
-import { PageNotFound } from "../app/pages/PageNotFound"
+import { MyTickets } from "../app/pages/MyTicketsPage"
+import { MyTicketsSold } from "../app/pages/MyTicketsSoldPage"
+import { PageNotFound } from "../app/pages/PageNotFoundPage"
 import PagarEntrada from "../app/pages/PayEntryPage"
-import PhoneCode from "../app/pages/PhoneCode"
+import PhoneCode from "../app/pages/PhoneCodePage"
 import VenderEntrada from "../app/pages/SellEntryPage"
 import Ticket from "../app/pages/TicketPage"
 import UserProfile from "../app/pages/UserProfilePage"
